shifted tests for diginorm over to new file
